backend/requirements.txt
fastapi
langchain_community
chromadb
pydantic
sentence-transformers
pypdf
python-multipart